Tweaked hostage system to hold kidnapper responsible for hostage's well-being (#50)
* Added variable for hostage to store kidnapper

Variable set on hostage player instance for later use (e.g. death handling)

* Evaluate hostage condition on gutting action

On "medGut" _unit's status as a possible hostage is evaluated and, if applicable, _medic's point loss or gain is redirected to and evaluated based on the original perpatrator.

* Evaluate victim's hostage status for score calculation

If the victim was being held hostage at the time of death the original kidnapper will be held responsible (both for good and bad) and points gain or less will be applied accordingly.