Update CfgBuildingLoot.hpp
I have fixed the annoying cfg.weapons error that throws in every BreakingPoint session. It was a simple "G" character that was added to two configs in the Heli Crash loot systems at some point. This error has existed for a long time now. This minor fix will remove that error throw.

It was..
BP_M40A3G
changed to...
BP_M40A3
It was...
BP_M24DesG
changed to...
BP_M24Des