Fixed errors
Wrong classname for base class C_man_polo_BP updated, fixing an error that throws every time a player joins a server.

Fixed two description related errors with the Worn Tire and Melee Hatchet.

Fixed two weapon description errors with the SVD and SVDK.

Improved the infoscreen.paa text from dark red to white and adjusted the size of the map name so the full map name shows on the diary when accessed.